Moved source code to its own directory and fixed getSelection
