+++ source/fileOps.c
@@ -1,4 +1,4 @@
-#include "./includes/fileOps.h"
+#include "../includes/fileOps.h"
 
 void writeDest(char* string, int mode){
 	FILE* writeTo = NULL;
