+++ source/helpers.c
@@ -1,4 +1,4 @@
-#include "./includes/helpers.h"
+#include "../includes/helpers.h"
 
 //printing an Error is more dynamic and more compact
 //this also means there is no table to shift
