+++ source/linkedList.c
@@ -1,4 +1,4 @@
-#include "./includes/linkedList.h"
+#include "../includes/linkedList.h"
 
 void addToList(Node_t** head, const char* string){
 	Node_t* newNode = malloc(sizeof(Node_t));
